requirements.txt
celery[redis]
certifi>=14.05.14
common-dibbs
django>=1.8,<1.9
django-jsonfield
django-periodically
django-states
djangorestframework>=3.5,<3.6
python_dateutil>=2.5.3
redis
setuptools>=21.0.0
six>=1.8.0
urllib3>=1.15.1
